+++ lib/highs/model.rb
@@ -2,7 +2,7 @@ module Highs
   class Model
     def initialize
       @ptr = FFI.Highs_create
-      ObjectSpace.define_finalizer(self, self.class.finalize(@ptr))
+      ObjectSpace.define_finalizer(@ptr, self.class.finalize(@ptr.to_i))
 
       check_status FFI.Highs_setBoolOptionValue(@ptr, "output_flag", 0)
     end
@@ -45,9 +45,9 @@ module Highs
       @ptr
     end
 
-    def self.finalize(ptr)
+    def self.finalize(addr)
       # must use proc instead of stabby lambda
-      proc { FFI.Highs_destroy(ptr) }
+      proc { FFI.Highs_destroy(addr) }
     end
 
     private